Fixed difficulty button main menu + improved player shooting time + added enemy archer unit

+++ Defense/Assets/Scripts/Data/BaseData.cs
@@ -7,7 +7,7 @@ public class BaseData : ScriptableObject {
 
 	[Header("Current")]
 	[SerializeField] private int coins = 0;
-	[SerializeField] private int maxHealth = 100;
+	[SerializeField] private float maxHealth = 100;
 	[SerializeField] private float damageOutput = 5;
 	[SerializeField] private float attackSpeed = 1;
 	[SerializeField] private float projectileSpeed = 1;
@@ -19,12 +19,12 @@ public class BaseData : ScriptableObject {
 	[SerializeField] private float beginProjectileSpeed;
 
 	public int Coins { get { return coins; } set { coins = value; } }
-	public int MaxHealth { get { return maxHealth; } set { maxHealth = value; } }
+	public float MaxHealth { get { return maxHealth; } set { maxHealth = value; } }
 	public float DamageOutput { get { return damageOutput; } set { damageOutput = value; } }
 	public float AttackSpeed { get { return attackSpeed; } set { attackSpeed = value; } }
 	public float ProjectileSpeed { get { return projectileSpeed; } set { projectileSpeed = value; } }
 
-	public int BeginMaxHealth { get { return beginMaxHealth; } }
+	public float BeginMaxHealth { get { return beginMaxHealth; } }
 	public float BeginDamageOutput { get { return beginDamageOutput; } }
 	public float BeginAttackSpeed { get { return beginAttackSpeed; } }
 	public float BeginProjectileSpeed { get { return beginProjectileSpeed; } }
